Set OVS max revalidator delay to 200ms (#1584)
The default value (500ms) causes a K8s upstream conformance test to fail
and may increase the risk of users reporting issues if they think
SessionAffinity is not implemented properly. Of course there is still a
delay before a learned flow is installed in the datapath flow cache and
takes effect, but the delay is at least smaller now. This delay cannot
be avoided when using the learn action.

Reducing the delay to 200ms may have an impact on performance with a
large number of datapath flows. Hopefully this impact is not
significant. There are other measures we could investigate if we notice
performance issues (e.g. reduce the max-idle delay for cached datapath
flows to reduce the size of the cache).

Fixes #1583