Change L2 forwarding flows for traffic to gateway and tunnel
Let the packets to the tunnel port go through L2ForwardingCalcTable
as well, to be consistent with other types of traffic.
Redirect packets to gateway and tunnel, as well as BUM packets to
conntrackCommitTable and bypass ingress NetworkPolicy enforcement.
Let l3DecTTLTable decrement TTL for packets from tunnel to local Pods.