Fix next table decision of l2ForwardingCalcTable
Before the commit, ofPort number is used to decide whether NetworkPolicy
ingress tables should be bypassed after l2ForwardingCalcTable - when
the output ofPort is DefaultTunOFPort or HostGatewayOFPort, the flow
will go to conntrackCommitTable and such bypass the ingress rule tables.
However, in the noEncap and networkPolicyOnly modes, the default tunnel
port is not created on the bridge, and DefaultTunOFPort can be used by
a Pod's OVS port, so the implementation might mistakenly bypass ingress
rule enforcement for traffic to the Pod.
This commit fixes the problem by letting the callers decide whether or
not to bypass ingress rule tables.