Fix in Traceflow e2e test
inter-Node Traceflow test should be skipped in Kind because of a know
issue when using Geneve with the userspace OVS datapath, However, the
logic for skipping was incorrectly added to the intra-Node test instead
of the inter-Node one.

The commit also removes comments describing the list of Traceflow test
cases: the comments were out-of-date and not convenient to keep in sync
with the actual test cases. The name of each test case should be a
sufficient description of what the test is doing.