Fix flaky upgrade e2e test
Due to netdev datapath bug, sometimes datapath flows are not flushed
after new openflows that change the actions are installed, causing
client1 to still be able to connect to the servers after after creating
a policy that disallows it. The test waits for 10 seconds so that the
datapath flows will expire.