Agent side changes for Flow Aggregator
We extended the DNS name support for flowCollectAddr configmap parameter.
If empty proto is given, we consider tcp as proto (default). If empty
empty port is given, we consider 4739 as default. This is to support the newly
deployed flow-aggregator service.
We resolve the flow aggregator IP using the static DNS
name: "flow-aggregator.flow-aggregator.svc.cluster.local".

In addition, antrea-agent base manifest has to be modified to set appropriate
DNS policy.