Necessary change in codegen image after update to Go 1.15 (#1634)
https://github.com/vmware-tanzu/antrea/pull/1420 updated the version of
Go used in the Antrea build infra from 1.13 to 1.15, including the
version of Go used in the antrea/codegen Docker image. However, an
updated codegen image was never pushed to the Dockerhub registry. It
turns out that an additional change was required in the image, caused by
a change of behavior in Go vendoring introduced in Go 1.14.

This is also the opportunity to update the version of mockgen (which
does not cause any change in the generated code).