Add e2e tests for flow aggregator
In this commit, we add e2e tests for the flow aggregator. For these
tests, we setup an environment where the flow exporters in antrea agents
export flow records to the flow aggregator. The flow aggregator will run
the collecting, aggregating jobs and will export records to the external
IPFIX collector.

The existing "flow exporter" tests (where the agents export to the
external collector directly) are replaced with "flow aggregator" tests,
as we assume that the aggregator is always deployed.

Also, we removed the configuration in the agents causing flow records to
only be exported from the source node since we now have the aggregator
to correlate and de-duplicate flows. Related commit and PR are b44fd78,
PR#1268.