Implement usage reporting (telemetry) in Antrea Controller
Usage reporting can be enabled (it is off by default) by setting
"enableUsageReporting" to "true" in the Controller configuration.

Whhen enabled, the Controller will periodically (every 24 hours) gather
some usage statistics (from the K8s APIs and its own internal state) and
send them as JSON to a web service (https://telemetry.antrea.io/report)
using a POST over HTTPS.

The collected data is fully anonymized. Because we need to identify
distinct reports as coming from the same cluster, we auto-generate a
fully random "cluster UUID" when the Controller is deployed to a cluster
for the first time, and persist it to a ConfigMap. The "cluster UUID" is
included in every report sent from that cluster.

See #1692 for more details