The main logic implementation of the flow aggregator (#1596)
This is the second PR of flow aggregator implementation, which adds the
main logic of flow aggregator, including collecting, intermediate, and
exporting process. It is based on a lastest version go-ipfix@v0.3.2.

Also mock interfaces of ipfix collecting and intermediate process are
added in pkg/agent/flowexporter/ipfix and this directory has been moved
to pkg/ipfix since both flow exporter and flow aggregator need it for
unit tests.