Update Emergency Tier's Priority to 50 (#1665)
This PR updates the priority of Tiers created by Antrea, and space them out evenly. The current Antrea generated Tier
priorities do not allow enough room for user defined Tiers to be created, for example Emergency Tier allows only 4 user
created Tiers with higher priority. Since it ultimately is up to the users as to which Tiers go atop and the purpose of
Antrea created Tiers is mainly that of convenience, we decided to distribute the priorities more evenly across available
priority space. For example, setting the priority of the top most Tier of Emergency to a value of 50 gives enough room for
admins to create their own Tier hierarchy without having to rely on Antrea created Tiers.