Add Namespace and Service name as labels using mutating webhook
Add a mutation webhook to Antrea Service which labels all
Namespaces and Services that are created with the following label:
  `antrea.io/metadata.name: <resourceName>`
Such labels can then be used by Antrea-native policies and K8s
NetworkPolicies to write policies which specifically targets
Namespaces by their name using the traditional labelSelector
mechanisms.

The webhook stomps the resources with this label on every CREATE
or UPDATE event such that it is guaranteed that no update of
resource removes the label.