Support building Antrea docker images for arm/v7 and arm64
* Download containernetworking plugins for correct architecture in the
  "base" image Dockerfile. Inspired by
  https://github.com/docker-library/golang/blob/master/Dockerfile-debian.template.
* Rename build_and_push.sh scripts to build.sh, and add command-line
  options to tweak their behavior. In particular add '--platform' to
  specifiy the target platform. Note that just providing '--platform
  arm64' will not let you build an arm64 image on an amd64 build
  server. However, some build servers can support multiple
  architectures. This is typically the case of arm64 servers, which can
  build images for both 32-bit and 64-bit arm target platforms.
* Update run-k8s-e2e-tests.sh to support arbitrary sonobuoy and
  conformance images, and to install correct sonobuoy binary based on
  the architecture.
* Cross-compile Antrea binaries for arm/v7 and arm64 as part of CI
  (Github actions) to ensure that arm builds are not broken.

See #450

We will use self-hosted arm64 runners to build and test the Antrea
Docker images for arm platforms. Note that for security reasons (using
self-hosted runners for public Github repositories is not recommended),
the images will be built and pushed from a separate private Github
respository for now. I believe Github is working on a solution to
improve security, which should enable us to move our arm Github
workflows to the main Antrea repository in the future.