Fix flaky test TestControllerWithDuplicatePodCIDR (#1723)
The test must wait for cache sync first, otherwise resource creation
events will be missing if the resources are created in-between list and
watch call of an informer. This is because fake clientset doesn't
support watching with resourceVersion. A watcher of fake clientset only
gets events that happen after the watcher is created.