antrea-agent supports overriding apiserver
antrea-agent relies either the following clientConnection or in-cluster
serviceaccount to configure the clientset that talks to apiserver.

```
clientConnection:
  kubeconfig: /home/ubuntu/.kube/config
```

For in-cluster configuration the apiserver host and port are injected
by kubelet, which is normally the apiserver's service vip (e.g.
10.96.0.1), which implies dependency upon kube-proxy. This new option
allows overriding them by `kubeAPIServerOverride`. It must be a host
string, a host:port pair, or a URL to the base of the apiserver.

For example:
```
kubeAPIServerOverride: 127.0.0.1:43305
```

Fixes #1684