Check the flow-restore-wait config before and after deleting it (#1730)
A rare case has been found that the "flow-restore-wait" config was still
true even though the delete call below was considered success. At the
moment we don't know if it's a race condition caused by "ovs-vsctl set
--no-wait" or a problem with OVSDB golang lib or OVSDB itself. To work
around it, we check if the config is true before deleting it and if it
is false after deleting it, and we will log warnings and retry a few
times if anything unexpected happens.