Add ClusterGroup API types (#1639)
Add ClusterGroup CRD which is responsible for collecting Pods and Namespaces across the cluster based on labelselectors defined in the ClusterGroup definition. It also allows setting an IPBlock (cannot be set with other Selectors) in the ClusterGroup. The purpose of a ClusterGroup is to allow grouping of resources and then be referenced in Antrea ClusterNetworkPolicies without having to add the same selectors in every ACNP when the group of resources are meant to be shared. This allows for greater sharing and decouples the job of reconciling effective group members from that of enforcing security policies.

This PR adds the following:

ClusterGroup API types
ClusterGroup CRD YAML
Controller to reconcile effective members of a ClusterGroup
controlplane group resource to be later used to retrieve realized members of a ClusterGroup
expose controlplane group via antrea-controller apis
Validation webhook to validate a ClusterGroupSpec