Conditionally pull base image for "make ubuntu" target (#1748)
It is not always appropriate to pull the base image when building the Antrea
Docker image with the "make ubuntu" target. In particular, when using a Docker
registry other than docker.io, we may not want to pull the base image. Other
targets already consider the "DOCKER_REGISTRY" variable and only pull the base
image if the variable is not set. We do the same for the "make ubuntu" target.
Other variables may be used in the future to make this decision if appropriate.