Extract and optimize group member calculation
Previously grouping member calculation was coupled with the
NetworkPolicyController, and there were two kinds of heavy calculation
in the process: For each event like Pod add, it needed to find all
affected groups by matching the new Pod's labels with all groups that
can potentially select it. If it was an update event, it even needed to
perform the process one more time for its old labels. Then it triggered
the sync process of the affected groups. For the sync process of a
group, it needed to find all Pods or ExternalEntities by matching its
selectors with all entities that can potentially match it. For each pair
of Pod and Group, they would be matched at least twice. Besides,
whenever the group was processed another time, it would again scan all
Pods.

The repeated calculation can be eliminated by caching the result of the
first matching process when processing Pod events. Then the matching
process will only occur when there's a new Pod or a new Group. Besides,
most Pods actually share same labels with others when they are managed
by same controller. The cache can be further optimized to maintain the
relationship between label set and label selector, making Pods that
having same labels and Groups that having same selectors share the
cache.

This patch introduces an interface which is responsible for group member
calculation and an implementation implementing it with the above
optimization. The interface is then consumed by NetworkPolicyController
and features like EndpointQuerier and ClusterGroup to retrieve
Pods/ExternalEntities selected by a given Group or Groups that select a
given Pod. Besides, other controllers that have the group logic like
EgressPolicy can consume it directly without having redundant code with
NetworkPolicyController.