Add Status subresource to ClusterGroup CRD (#1778)
Add status sub resource to ClusterGroup CRD. The status will store the conditions associated with a CG.
Currently, supported values for status.conditions are conditions of type GroupMembersComputed. The condition status
of the GroupMembersComputed will store whether the CG has computed its members. Valid values for the condition
status of GroupMembersComputed type are "True" and "False". The condition status will be updated by the controller to
"True" once group members have been calculated and updated to the corresponding internal Group.