Fortify the bandwidth e2e test of flow aggregator
Currently, bandwith test is disabled in flow aggregator e2e test because
it's flaky especially for intra-node flows. We found the root cause of
it is the first and final records exported by flow exporter don't have
desired octectDeltaCount because iperf traffic may not fill the entire
export interval of the first and final records.

So in this commit, we check bandwidth using
octetDeltaCountFromSourceNode for the records in the middle and
calculate the average bandwidth using the octetTotalCountFromSourceNode
, then compares with the iperf bandwidth. The difference should be lower
than 15%.

Signed-off-by: Yongming Ding <dyongming@vmware.com>