Service annotation based pod selection for NodePortLocal (#1743)
Currently the NPLController in antrea-agent watches for all pods on the
node, and exposes all pods to the external network using
nodeip:nodeport. This PR introduces pod selection/filtering for
NodePortLocal based on an annotation to be provided in the Service
objects. The Service objects can be of any type but NodePort. The
proposed annotation NPLController will watch for is:
nodeportlocal.antrea.io/enabled: "true"

The value "true" is case sensitive here. This makes the selection and
implementation of NodePortLocal much more controlled.

Another key update as part of this is to avoid exporting Pod/Service
informers from the NPLController. All the eventHandler functions
(enqueuePod/enqueueSvc) have also been made local to the controller and
are not being exported anymore. This involved moving the
AddEventHandlerWithResyncPeriod and AddIndexers functions inside
NewNPLController(...) as well, from
pkg/agent/nodeportlocal/npl_agent_init.go to
pkg/agent/nodeportlocal/k8s/controller.go