Fix: add Endpoints references counter
In current implementation, if an Endpoint relates to multiple
Services, once a Service deleted, flows of the Endpoint will
also be deleted. This commit adds a Endpoint references counter
which helps to determine if a Endpoint should be removed from
data path.

This commit also contains log statements to improve debug
ability.

Signed-off-by: Weiqiang Tang <weiqiangt@vmware.com>