Fix issues with NPL unit tests
The unit tests were fundamentally broken: they were not independent from
each other and assumed a specific execution. It was not possible to run
a single unit test in isolation with the others. The tests were also
modifying the environment, with no cleanup logic.

This commit also tries to improve readability of the tests and reduce
their verbosity.