Add ClusterGroup reference in ClusterNetworkPolicy rules (#1734)
Allow ClusterNetworkPolicies to select ClusterGroups in to/from rules instead of standalone labelSelectors.

ClusterNetworkPolicy rules can now be set with a list of ClusterGroups in group for ingress and egress rules
instead of setting stand alone selectors. These ClusterGroup references in rules will be converted to corresponding
AddressGroups with the UID matching that of it's ClusterGroup. Group members calculated by the internal Group
corresponding to the ClusterGroup will be copied in to these AddressGroups and Agents will continue to consume these
source/destinations in the form of AddressGroup messages.

A ClusterGroup must exist before being referenced by ACNP. Similarly, a CG cannot be deleted if it is referenced by at least one ACNP.