Add liveness probe to Antrea Controller
This is motivated by #1837, in which the Antrea Controller got in a bad
state (presubmably because of a bug in the apiserver library), and
didn't recover. The endpoint was marked as non-Ready for the Antrea
Service, which means the APIService became broken, since the Antrea
Service has a single Endpoint (single Controller replica). With the
liveness probe, the Antrea Controller would have been restarted by
kubelet and the APIService would hopefully have recovered.

The /healthz endpoint is deprecated since K8s v1.16, so we switch to
/readyz and /livez.