Cyclonus netpol test case generator (#1765)
This is proof-of-concept work for #1761. The
https://github.com/mattfenwick/cyclonus project is used to generate
NetworkPolicy tests cases and run them. Test cases are generated
automatically, but this is not real "fuzz-tesing" as we always generate
the same set of test cases.

Cyclonus is run as a daily CRON Job using Github actions. The set of
tests cases is limited for now but can be expanded in the future.