Fix: add Endpoints references counter (#1815)
In the current implementation, if an Endpoint relates to multiple
Services, once a Service deleted, flows of the Endpoint will
also be deleted. This commit adds an Endpoint references counter
which helps to determine if an Endpoint should be removed from
data path.

This commit also adds logs to improve debuggability.

Fixes #1835.
Signed-off-by: Weiqiang Tang weiqiangt@vmware.com
Co-authored-by: Antonin Bas <antonin.bas@gmail.com>