Add ClusterGroup reference in ClusterNetworkPolicy AppliedTo (#1750)
ClusterNetworkPolicy spec/rules can now be set with a list of ClusterGroups in appliedToGroups in spec or ingress rules
and egress rules instead of setting stand alone selectors. These ClusterGroup references in rules will be converted to corresponding control plane AppliedToGroups with the UID matching that of it's ClusterGroup.Group members calculated
by the internal Group corresponding to the ClusterGroup will be used to determine workloads and set in AppliedToGroup
and Agents will continue to consume these appliedtos in the form of AppliedToGroup messages.