Handle restartability for Node Port Local (#1758)
When the Antrea Agent restarts, instead of removing the NPL annotations
from all Pods before starting the NPL Controller, we reconcile existing
Pod annotations with iptables rules using iptables-restore. We also
re-build the port table / port cache. We then start running the NPL
Controller. If the Pod NPL annotations are no longer up-to-date
(e.g. because a Service selector changed), it will be handled when the
Controller is running, just like any normal update.

This is a follow-up work item from #1459.