Ensure that "antctl version" always outputs the client version
Update antctl command framework to support outputting information to the
user even when a client request to the server (Agent or Controller)
fails. We leverage this framework to ensure that "antctl version" always
outputs the client version.

To test this functionality, we introduce an interface for the "client"
part of antctl, which is in charge of sending a request to the
appropriate server. We mock this interface so it can be used in unit
tests. Although in the Antrea codebase, we typically place mocks in a
sub-package called "testing", it is not possible here, due to the very
flat nature of the antctl package, and to the dependencies between
objects. For example, "commandDefinition" depends on "client" and
"client" depends on "commandDefinition". Changing this structure would
require a large code refactor, so instead we adapt the code generation
framework to exceptionally tolerate placing mocks directly in the
package.

Fixes #1872