move go-ipfix to v0.4.5 (#1890)
This commit adds changes in go-ipfix v0.4.5. It modifies the set
interface: add PrepareSet and ResetSet methods to support the re-use of the set to reduce the allocations
and enables to write more unit tests to increase the coverage.
It also simplifies the input consumption for
exporting and collecting process.