Trigger Docker multi-arch build on pushes and releases
When the main branch is updated, or a new relese tag is creted, we build
and push the amd64 Antrea Docker image (renamed from
antrea/antrea-ubuntu to antrea/antrea-ubuntu-amd64), then trigger a
workflow in the private vmware-tanzu/antrea-build-infra repository which
will build the Antrea ARM images (antrea/antrea-ubuntu-arm and
antrea/antrea-ubuntu-arm64) as well as the multi-arch manifest list
(antrea/antrea-ubuntu).

As a reminder, vmware-tanzu/antrea-build-infra is private (maintainer
access only) because of security issues when using self-hosted runners
(in our case, arm64 machines) in a public repo. When / if these security
issues are addressed by Github, we will merge
vmware-tanzu/antrea-build-infra into vmware-tanzu/antrea.

See #450