Revert some changes
The ConfigMap is antrea=cluster-identity. At the moment it only stores
a UUID value. The UUID is meant to be generated by the Antrea
Controller, and there is no longer a way for a user to provide it
through config. In theory, a user can populate the ConfigMap directly
before starting the Controller, and Antrea will use that value (it will
not be overridden). There is no longer a cluster name.

Also added an end-to-end test.