Ensure Github workflows use up-to-date base Docker images
Relying on a CRON job to update the Antrea base images has proven
sub-optimal: we sometimes push base images manually which do not match
the checked-in Dockerfiles and CI tests run for PRs which update the
Dockerfiles ignore these updates (making the tests worthless).

We now ensure that Github workflows always build the base images befgore
building the Antrea image, thanks to a new helper script. By relying on
Docker caching (using the Dockerhub registry as the cache), we ensure
that build times are not increased: in the absence of any change, we
only add a handful of seconds to the build time.

For now, we only update CI jobs run as Github workflow. Once this is
merged, we should consider doing the same for Jenkins scripts. We could
add support for DOCKER_REGISTRY to the new helper script
(hack/build-antrea-ubuntu-all.sh).

One question that we could ask now is whether these base images are even
necessary: if caching works well, using one large Dockerfile should be
just as fast, while simplifying build architecture. This is something we
may want to revisit in the future. Maybe using base images only makes
sense if we are going to share them across multiple images.

See #1540