Change Mellanox CI logs URL to point to the HTML logs directly
Previously the CI used to point to the Jenkins job console as its
context URL, and the end of that, there was a link to an HTML server
that contained the logs. As a part of logs improvement, a change to
directly point to the HTML server with the console being uploaded to
it was adopted. This was done for ease of accessing the logs, and
to avoid slow uploading times of uploading the logs from the internal
Jenkins to the Public Jenkins.