Fix Docker build when main branch is updated
The build/images/base/build.sh script was incorrectly using the
OVS_VERSION env variable as the tag for the intermediate cni-binaries
image. Instead, we should be using the version of the CNI binaries
packaged in the Docker image.

This is currently causing the CI build on the main branch to fail,
preventing the antrea/antrea-ubuntu:latest Docker image to be updated
automatically.