Store OVS version in a single place
Instead of repeating the default OVS version (currently 2.14.0) all over
the place, and of requiring that the OVS_VERSION environment variable be
set for build scripts, we use a file (build/images/deps/ovs-version) to
store the version number. Updating this file will change the OVS version
used for the entire build process. The only drawback is that we need to
make the Makefile a bit more verbose, since "docker build" needs to be
invoked with the appropriate build arg.

We do the same thing for the CNI binaries version.

Additionally, we fix an issue with hack/build-antrea-ubuntu-all.sh: when
invoking the script with "--pull" and without "--push", the
antrea/openvswitch image built locally would not be used when building
the antrea/base-ubuntu image.