Packet-in rate limiting with OF Meter
Add OF meter to implement rate-limiting for packet-in messages.

1. Add meter entries during initialization: one for Traceflow packets
   and one for other (NetworkPolicy-related) packets.
2. While building the flow that will trigger packet-in, use the meter
   action.
3. Update libOpenflow and ofnet version, to get meter programming
   support.

Since Windows OVS doesn't support OF meters, we skip OF meter related
operations for now. On Linux, for the OVS kernel datapath, kernel
version 4.18 is required for meter support (should be 4.15, but is 4.18
in practice because of an implementation bug): we add a check and
disable meters if the Linux kernel is not recent enough. This is to
avoid increasing the minimum kernel version requirement for Antrea, at
least for now.

Some issues were observed during Kind testing: some stale control socket
files (*.ctl) were kept under /var/run/openvswitch after restarts,
causing issues when running ovs-apptcl commands. It's hard to see how
this could be related specifically to this change, but in order to
work around the issue, we remove all stale .ctl files when restarting
OVS.

Fixes #2069

Co-authored-by: Antonin Bas <abas@vmware.com>

Signed-off-by: wgrayson <wgrayson@vmware.com>
Signed-off-by: Antonin Bas <abas@vmware.com>