Re-organize CHANGELOG by making it a directory
This lets us have a different CHANGELOG file for each minor release
train (same as K8s), which makes it easier to navigate changes.

When publishing the CHANGELOG for a new minor release, all the changes
since the first version of the previous minor release should be
included. This removes any ambiguity as to which of the bug fix releases
is "included" in the minor release.

We also introduce a new workflow which will take care of automatically
updating the appropriate CHANGELOG file on the main branch whenever a
new version of Antrea (minor or bug fix) is released. This will ensure
that the CHANGELOG on the main branch is always up-to-date, which has
not been the case in the past. The workflow cannot commit to the main
branch directly (it is protected), but it will open a PR automatically.

See #1532